Refactored to simplify iteration in Changes

+++ lib/reenrb/change.rb
@@ -32,16 +32,16 @@ module Reenrb
 
     def compare
       if original == requested
-        :none
+        CHANGE::NONE
       elsif requested.start_with? "-"
-        :delete
+        CHANGE::DELETE
       else
-        :rename
+        CHANGE::RENAME
       end
     end
 
     def consider
-      if request_full_directory? && request_delete?
+      if request_full_dir? && request_delete?
         @status = STATUS::REJECTED
         @reason = "Directories with files cannot be changed"
       else
@@ -53,9 +53,9 @@ module Reenrb
       return(self) if executed_or_rejected?
 
       case @change
-      when :rename
+      when CHANGE::RENAME
         File.rename(@original, @requested)
-      when :delete
+      when CHANGE::DELETE
         File.delete(@original)
       end
 
@@ -65,15 +65,23 @@ module Reenrb
 
     # Predicates
 
-    def request_directory? = Dir.exist?(@original)
+    def request_dir? = Dir.exist?(@original)
+
+    def request_nothing? = @change == CHANGE::NONE
+
+    def request_rename? = @change == CHANGE::RENAME
 
     def request_delete? = @change == CHANGE::DELETE
 
-    def request_empty_directory? = Dir.empty?(@original)
+    def request_empty_dir? = Dir.empty?(@original)
+
+    def accepted? = @status == STATUS::ACCEPTED
 
     def executed? = @status == STATUS::EXECUTED
 
-    def request_full_directory? = request_directory? && !request_empty_directory?
+    def rejected? = @status == STATUS::REJECTED
+
+    def request_full_dir? = request_dir? && !request_empty_dir?
 
     def executed_or_rejected? = %i[executed rejected].include?(@status)
 
@@ -82,7 +90,7 @@ module Reenrb
     def to_s
       file_desc =
         case @change
-        when :rename
+        when CHANGE::RENAME
           "#{@original} -> #{@requested}"
         else
           @original
